requirements.txt
aiogram==3.13.0
finvizfinance==1.0.1
yfinance==0.2.43
python-dotenv==0.20.0
tradingview-screener==2.5.0
edgartools==2.29.1